move format & parsing of timestamps to re_format and improve it a little bit
