+++ crates/utils/re_format/src/lib.rs
@@ -6,7 +6,7 @@ mod time;
 
 use std::{cmp::PartialOrd, fmt::Display};
 
-pub use time::next_grid_tick_magnitude_ns;
+pub use time::{format_timestamp_seconds, next_grid_tick_magnitude_ns, parse_timestamp_seconds};
 
 // --- Numbers ---
