+++ crates/utils/re_format/src/time.rs
@@ -18,3 +18,48 @@ pub fn next_grid_tick_magnitude_ns(spacing_ns: i64) -> i64 {
         spacing_ns.checked_mul(10).unwrap_or(spacing_ns) // multiple of ten days
     }
 }
+
+/// Formats a timestamp in seconds to a string.
+///
+/// This is meant for a relatively stable display of timestamps that are typically in minutes
+/// where we still care about fractional seconds.
+pub fn format_timestamp_seconds(timestamp_seconds: f64) -> String {
+    let n = timestamp_seconds as i32;
+    let hours = n / (60 * 60);
+    let mins = (n / 60) % 60;
+    let secs_int = n % 60;
+    let milliseconds = (timestamp_seconds.fract() * 1000.0) as u32;
+
+    if hours > 0 {
+        format!("{hours:02}:{mins:02}:{secs_int:02}.{milliseconds:03}")
+    } else {
+        format!("{mins:02}:{secs_int:02}.{milliseconds:03}")
+    }
+    // Not showing the minutes at all makes it too unclear what format this timestamp is in.
+    // So let's not further strip this down.
+}
+
+/// Parses seconds from a string.
+///
+/// Supports:
+/// * fractional seconds
+/// * minutes:[fractional seconds]
+/// * hours:minutes:[fractional seconds]
+pub fn parse_timestamp_seconds(s: &str) -> Option<f64> {
+    let parts: Vec<&str> = s.split(':').collect();
+    match parts.len() {
+        1 => parts[0].parse::<f64>().ok(),
+        2 => {
+            let minutes = parts[0].parse::<i32>().ok()?;
+            let seconds = parts[1].parse::<f64>().ok()?;
+            Some((minutes * 60) as f64 + seconds)
+        }
+        3 => {
+            let hours = parts[0].parse::<i32>().ok()?;
+            let minutes = parts[0].parse::<i32>().ok()?;
+            let seconds = parts[1].parse::<f64>().ok()?;
+            Some(((hours * 60 + minutes) * 60) as f64 + seconds)
+        }
+        _ => None,
+    }
+}
