+++ crates/viewer/re_component_ui/src/lib.rs
@@ -137,7 +137,7 @@ pub fn create_component_ui_registry() -> re_viewer_context::ComponentUiRegistry 
 
     registry.add_singleline_edit_or_view(entity_path::edit_or_view_entity_path);
 
-    registry.add_singleline_edit_or_view(video_timestamp::edit_or_view_entity_path);
+    registry.add_singleline_edit_or_view(video_timestamp::edit_or_view_timestamp);
 
     registry
 }
